refactor: found better home for params
